pyproject.toml
[project]
name = "nonebot-nix"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nb-cli>=1.4.2",
    "nonebot-adapter-qq>=1.6.2",
    "nonebot-plugin-maimai-updater>=0.1.0",
    "nonebot-plugin-orm[default]>=0.7.7",
    "nonebot-plugin-send-anything-anywhere>=0.7.1",
    "nonebot2[fastapi,httpx]>=2.4.2",
]

[tool.nonebot]
adapters = [{ name = "QQ", module_name = "nonebot.adapters.qq" }]
plugins = ["nonebot_plugin_maimai_updater"]
plugin_dirs = []
builtin_plugins = []

[project.scripts]
nonebot-run = "bot:main"